#7:  404 route with custom pages/404.tsx
